Add files via upload
This currency converter application is built using Java Servlets and JSP (JavaServer Pages). The application allows users to convert an amount from one currency to another using real-time exchange rates fetched from an external API. The HTML form (index.jsp) provides a simple user interface for inputting the amount and selecting the currencies to convert from and to. The CurrencyConverter servlet class handles the form submission. It processes the input data, fetches the current exchange rate from the CurrencyFreaks API, performs the conversion, and forwards the result to a JSP page for display. The Result.jsp page displays the conversion result and provides a link to perform another conversion.